backend/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
black==22.12.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.3
django-environ==0.9.0
djangorestframework==3.14.0
gitdb==4.0.9
GitPython==3.1.30
idna==3.4
mypy-extensions==0.4.3
oauthlib==3.2.2
codinasion-tools
pathspec==0.10.3
platformdirs==2.6.0
pycodestyle==2.9.1
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
smmap==5.0.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tweepy==4.11.0
urllib3==1.26.12
